# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "dovecote",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
      "SendResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "channel": {
            "type": "string"
          },
          "messageId": {
            "type": "string"
          },
          "detail": {
            "type": "object",
            "properties": {
              "text": {
                "type": "string"
              },
              "chatId": {
                "type": "string"
              }
            }
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "channel"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "error_description": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "error_description"
        ]
      },
      "NotifyRequest": {
        "type": "object",
        "properties": {
          "channel": {
            "type": "string"
          },
          "content": {
            "type": "object",
            "properties": {
              "text": {
                "type": "string"
              },
              "embed": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string"
                  },
                  "timestamp": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "color": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 16777215
                  },
                  "footer": {
                    "type": "object",
                    "properties": {
                      "text": {
                        "type": "string"
                      },
                      "icon_url": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "text"
                    ]
                  },
                  "image": {
                    "type": "object",
                    "properties": {
                      "url": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "url"
                    ]
                  },
                  "thumbnail": {
                    "type": "object",
                    "properties": {
                      "url": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "url"
                    ]
                  },
                  "author": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "url": {
                        "type": "string"
                      },
                      "icon_url": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "name"
                    ]
                  },
                  "fields": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "value": {
                          "type": "string"
                        },
                        "inline": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "name",
                        "value"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "required": [
          "channel",
          "content"
        ]
      },
...
      "ChannelsList": {
        "type": "object",
        "properties": {
          "channels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChannelConfig"
            }
          }
        },
        "required": [
          "channels"
        ]
      },
      "EnvReadResponse": {
        "type": "object",
        "properties": {
          "profile": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "profile",
          "value"
        ]
      },
...
      "TokenIssueRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "minLength": 1
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 1
          },
          "label": {
            "type": "string",
            "maxLength": 64
          },
          "expiresIn": {
            "type": "string",
            "enum": [
              "7d",
              "30d",
              "60d",
              "90d"
            ]
          }
        },
        "required": [
          "userId",
          "scopes"
        ]
      },
      "RevokeResponse": {
        "type": "object",
        "properties": {
          "revoked": {
            "type": "boolean"
          },
          "tokenId": {
            "type": "string"
          },
          "notice": {
            "type": "string"
          }
        },
        "required": [
          "revoked",
          "tokenId",
          "notice"
        ]
      },
...
    },
...
  },
  "paths": {
    "/v1/notify": {
      "post": {
        "description": "Send a notification to the configured channel. Requires scope `dovecote:notify`.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification dispatched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendResult"
                }
              }
            }
          },
          "400": {
            "description": "Request body failed zod validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Authorization header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope for this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/channels": {
      "get": {
        "description": "List configured notification channels. Requires scope `dovecote:read`.",
        "responses": {
          "200": {
            "description": "Channels listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelsList"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Authorization header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope for this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/env/{profile}": {
      "get": {
        "description": "Read an env profile value. Requires scope `dovecote:env:read`.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "pattern": "^[a-zA-Z0-9_-]+$",
              "example": "dev"
            },
            "required": true,
            "name": "profile",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Env profile resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvReadResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Authorization header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope for this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tokens": {
      "post": {
        "description": "Issue a new API token. Requires scope `dovecote:admin`.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenIssueRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenIssueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request body failed zod validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Authorization header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope for this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — Retry-After header indicates retry window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Admin token revocation pressure circuit open",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tokens/{tokenId}": {
      "delete": {
        "description": "Revoke an API token. Requires scope `dovecote:admin`.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "example": "tkn_01HZ..."
            },
            "required": true,
            "name": "tokenId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Token revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Authorization header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope for this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — Retry-After header indicates retry window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Admin token revocation pressure circuit open",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/tokens/{tokenId}/renew": {
      "post": {
        "description": "Renew a dvct_* token. Self-renew (own tokenId) requires no admin scope; admin can renew any tokenId.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "example": "tkn_01HZ..."
            },
            "required": true,
            "name": "tokenId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRenewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token renewed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenIssueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request body failed zod validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Authorization header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope for this resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded — Retry-After header indicates retry window",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Admin token revocation pressure circuit open",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}